+++ src/app/layout.tsx
@@ -1,9 +1,9 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { ReactNode } from 'react';
-import './globals.css';
 import { Header, Modal, NavbarTop } from '@/components';
 import { ModalContextProvider } from '@/contexts';
+import './globals.css';
 
 export const metadata: Metadata = {
   title: 'Create Next App',
