+++ src/app/loading.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import React, { ReactNode } from 'react';
+import { ReactNode } from 'react';
 import styles from './App.module.css';
 
 const Loading = (): ReactNode => {
