+++ src/components/Button/index.ts
@@ -1,3 +1,1 @@
-import Button from './Button';
-
-export default Button;
+export { default as Button } from './Button';
