+++ src/components/Collapse/Content.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
 import { ReactNode } from 'react';
-import { useCollapseContext } from './Collapse';
+import useCollapseContext from './useCollapseContext';
 
 type Props = {
   children: ReactNode;
