+++ src/components/Collapse/index.ts
@@ -1,3 +1,1 @@
-import Collapse from './Collapse';
-
-export default Collapse;
+export { default as Collapse } from './Collapse';
