+++ src/components/Header/index.ts
@@ -1,3 +1,1 @@
-import Header from './Header';
-
-export default Header;
+export { default as Header } from './Header';
