+++ src/components/Modal/index.ts
@@ -1,3 +1,1 @@
-import Modal from './Modal';
-
-export default Modal;
+export { default as Modal } from './Modal';
